Extract defaults to separate module
