+++ draftjs_exporter/html.py
@@ -1,7 +1,7 @@
 from __future__ import absolute_import, unicode_literals
 
 from draftjs_exporter.command import Command
-from draftjs_exporter.constants import DEFAULT_BLOCK_MAP, DEFAULT_STYLE_MAP
+from draftjs_exporter.defaults import BLOCK_MAP, STYLE_MAP
 from draftjs_exporter.entity_state import EntityState
 from draftjs_exporter.style_state import StyleState
 from draftjs_exporter.wrapper_state import WrapperState
@@ -14,8 +14,8 @@ class HTML():
     """
     def __init__(self, config={}):
         self.entity_decorators = config.get('entity_decorators', {})
-        self.wrapper_state = WrapperState(config.get('block_map', DEFAULT_BLOCK_MAP))
-        self.style_state = StyleState(config.get('style_map', DEFAULT_STYLE_MAP))
+        self.wrapper_state = WrapperState(config.get('block_map', BLOCK_MAP))
+        self.style_state = StyleState(config.get('style_map', STYLE_MAP))
 
     def call(self, content_state):
         entity_map = content_state.get('entityMap', {})
